backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
executing==2.2.1
Flask==3.1.2
Flask-Cors==4.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==9.7.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
openai==2.7.1
parso==0.8.5
pillow==12.0.0
prompt_toolkit==3.0.52
pure_eval==0.2.3
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyMuPDF==1.26.5
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
stack-data==0.6.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
